Manage Listings

+++ routes/web.php
@@ -18,25 +18,28 @@ use Illuminate\Support\Facades\Route;
 
 Route::get('/', [ListingsController::class, 'index']);
 
-Route::get('/create', [ListingsController::class, 'create']);
+Route::get('/create', [ListingsController::class, 'create'])->middleware('auth');
 
-Route::post('/listings', [ListingsController::class, 'store']);
+Route::post('/listings', [ListingsController::class, 'store'])->middleware('auth');
 
-Route::get('/listins/{listing}/edit', [ListingsController::class, 'edit']);
+Route::get('/listins/{listing}/edit', [ListingsController::class, 'edit'])->middleware('auth');
 
-Route::put('/listings/{listing}', [ListingsController::class, 'update']);
+Route::put('/listings/{listing}', [ListingsController::class, 'update'])->middleware('auth');
 
-Route::delete('/listings/{listing}', [ListingsController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingsController::class, 'destroy'])->middleware('auth');
+
+Route::get('listings/manage', [ListingsController::class, 'manage']);
 
 Route::get('/listings/{listing}', [ListingsController::class, 'show']);
 
 
-Route::get('/register', [UsersController::class, 'register']);
 
-Route::post('users/logout', [UsersController::class, 'logout']);
+Route::get('/register', [UsersController::class, 'register'])->middleware('guest');
+
+Route::post('users/logout', [UsersController::class, 'logout'])->middleware('auth');
 
 Route::post('/users', [UsersController::class, 'store']);
 
-Route::get('/login', [UsersController::class, 'login']);
+Route::get('/login', [UsersController::class, 'login'])->name('login')->middleware('guest');
 
 Route::post('/users/authenticate', [UsersController::class, 'authenticate']);
